refactor: create index rollup for FlexBox family
