+++ src/index.ts
@@ -28,6 +28,7 @@ export * from './Components/Dropdowns/Family/DropdownMenu'
 export * from './Components/EmptyState/EmptyState'
 export * from './Components/EmptyState/EmptyStateSubText'
 export * from './Components/EmptyState/EmptyStateText'
+export * from './Components/FlexBox'
 export * from './Components/FlexBox/FlexBox'
 export * from './Components/FlexBox/FlexBoxChild'
 export * from './Components/Form/Form'
